Files for user statistics

+++ Frontend/AndroidExample/app/src/main/java/com/example/androidexample/UserStats.java
@@ -4,17 +4,17 @@ public class UserStats {
 
     private int id;
 
-    private double totalAnswered;
+    private int totalAnswered;
 
-    private double totalCorrect;
+    private int totalCorrect;
 
-    private double totalIncorrect;
+    private int totalIncorrect;
 
     private int winStreak;
 
-    private double wins;
+    private int wins;
 
-    private double losses;
+    private int losses;
 
     private int questionsSumbitted;
 
@@ -56,27 +56,27 @@ public class UserStats {
         this.id = id;
     }
 
-    public double getTotalAnswered() {
+    public int getTotalAnswered() {
         return totalAnswered;
     }
 
-    public void setTotalAnswered(double totalAnswered) {
+    public void setTotalAnswered(int totalAnswered) {
         this.totalAnswered = totalAnswered;
     }
 
-    public double getTotalCorrect() {
+    public int getTotalCorrect() {
         return totalCorrect;
     }
 
-    public void setTotalCorrect(double totalCorrect) {
+    public void setTotalCorrect(int totalCorrect) {
         this.totalCorrect = totalCorrect;
     }
 
-    public double getTotalIncorrect() {
+    public int getTotalIncorrect() {
         return totalIncorrect;
     }
 
-    public void setTotalIncorrect(double totalIncorrect) {
+    public void setTotalIncorrect(int totalIncorrect) {
         this.totalIncorrect = totalIncorrect;
     }
 
@@ -112,19 +112,19 @@ public class UserStats {
         this.gamesPlayed = gamesPlayed;
     }
 
-    public double getLosses() {
+    public int getLosses() {
         return losses;
     }
 
-    public void setLosses(double losses) {
+    public void setLosses(int losses) {
         this.losses = losses;
     }
 
-    public double getWins() {
+    public int getWins() {
         return wins;
     }
 
-    public void setWins(double wins) {
+    public void setWins(int wins) {
         this.wins = wins;
     }
 }
